+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## TBD
+
+* Add warning in BAGP for React Native when running via react-native CLI
+  [#357](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/357)
+
 ## 5.7.0 (2021-01-13)
 
 * Support AGP 4.2